refactor: rename categories dir
