+++ src/category/dtos/batch-update-category.dto.ts
