+++ src/category/entities/category.entity.ts
